requirements.txt
rtmidi2
pyserial